Move logic to service to make it testable.
